requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pymupdf==1.24.9
python-docx==1.1.2
langchain==0.3.0
langchain-huggingface==0.1.0
langchain-community==0.3.0
chromadb==0.5.5
sentence-transformers==3.1.0
huggingface-hub==0.25.0
pandas==2.2.2
python-dotenv==1.0.1
